Tweak tilemap and level_size components

+++ src/components/ctx/level_size.hpp
@@ -1,11 +1,13 @@
 #pragma once
 
+#include "common/ints.hpp"
+
 namespace wanderer::ctx {
 
 struct level_size final
 {
-  int row_count{};  ///< The number of rows in the level.
-  int col_count{};  ///< The number of columns in the level.
+  uint64 row_count{};  ///< The number of rows in the level.
+  uint64 col_count{};  ///< The number of columns in the level.
 };
 
 }  // namespace wanderer::ctx
